+++ groovy/globals/CoolantGases.groovy
@@ -1,0 +1,15 @@
+package globals
+
+class CoolantGases {
+    String name
+    String byproduct
+    int amount
+    int duration
+
+    CoolantGases(name, byproduct, amount, duration) {
+        this.name = name
+        this.byproduct = byproduct
+        this.amount = amount
+        this.duration = duration
+    }
+}
