+++ groovy/postInit/chemistry/organic_chemistry/Surfactants.groovy
@@ -90,7 +90,7 @@ BR.recipeBuilder()
 MIXER.recipeBuilder()
     .fluidInputs(fluid('benzenesulfonic_acid') * 1000)
     .fluidInputs(fluid('sodium_hydroxide_solution') * 1000)
-    .outputs(metaitem('dustSodiumBenzeneSulfonate'))
+    .outputs(metaitem('dustSodiumBenzenesulfonate') * 16)
     .fluidOutputs(fluid('water') * 1000)
     .duration(100)
     .EUt(VA[MV])
