+++ groovy/postInit/chemistry/organic_chemistry/polymers/thermosets/RubberChain.groovy
@@ -315,9 +315,9 @@ MIXER.recipeBuilder()
 
 MIXER.recipeBuilder()
     .fluidInputs(fluid('butadiene') * 2880)
-    .fluidInputs(fluid('acrylonitrile') * 1008)
+    .fluidInputs(fluid('acrylonitrile') * 1000)
     .fluidInputs(fluid('water') * 5000)
-    .inputs(ore('dustSodiumBenzeneSulfonate') * 2)
+    .inputs(ore('dustSodiumBenzenesulfonate') * 16)
     .fluidOutputs(fluid('nitrile_emulsion') * 9000)
     .EUt(VA[MV])
     .duration(30)
@@ -325,7 +325,7 @@ MIXER.recipeBuilder()
 
 MIXER.recipeBuilder()
     .fluidInputs(fluid('hydrogen_peroxide') * 500)
-    .inputs(ore('iron_sulfate') * 1)
+    .inputs(ore('dustIronSulfate') * 3)
     .fluidOutputs(fluid('fenton_reagent') * 500)
     .EUt(VA[LV])
     .duration(30)
@@ -335,14 +335,13 @@ POLYMERIZATION_TANK.recipeBuilder()
     .fluidInputs(fluid('nitrile_emulsion') * 9000)
     .fluidInputs(fluid('fenton_reagent') * 500)
     .inputs(ore('hydroquinone') * 7)
-    .fluidOutputs('polymerized_nitrile_emulsion' * 10000)
+    .fluidOutputs(fluid('polymerized_nitrile_emulsion') * 10000)
     .EUt(VA[MV])
     .duration(500)
     .buildAndRegister()
 
-DT.recipeBuilder()
-    .fluidInputs(fluid('dense_steam') * 10000)
-    .fluidInputs('polymerized_nitrile_emulsion' * 5000)
+HIGH_TEMP_DT.recipeBuilder()
+    .fluidInputs(fluid('polymerized_nitrile_emulsion') * 5000)
     .fluidOutputs(fluid('butadiene') * 800)
     .fluidOutputs(fluid('acrylonitrile') * 200)
     .fluidOutputs(fluid('nitrile_latex') * 4000)
@@ -362,7 +361,7 @@ COAGULATOR.recipeBuilder()
 MIXER.recipeBuilder() // washing out
     .fluidInputs(fluid('water') * 1000)
     .inputs(ore('dustImpureNitrileRubber') * 4)
-    .outputs(metaitem('dustHydratedNitrileRubber'))
+    .outputs(metaitem('dustWashedNitrileRubber'))
     .fluidOutputs(fluid('sodium_benzenesulfonate_solution') * 500)
     .duration(40)
     .EUt(VA[LV])
@@ -370,13 +369,13 @@ MIXER.recipeBuilder() // washing out
 
 DISTILLERY.recipeBuilder()
     .fluidInputs(fluid('sodium_benzenesulfonate_solution') * 1000)
-    .outputs(ore('dustSodiumBenzeneSulfonate'))
+    .outputs(metaitem('dustSodiumBenzenesulfonate') * 16)
     .EUt(VA[ULV])
     .duration(10)
     .buildAndRegister()
 
 DRYER.recipeBuilder()
-    .inputs(ore('dustHydratedNitrileRubber') * 10)
+    .inputs(ore('dustWashedNitrileRubber') * 10)
     .outputs(metaitem('dustNitrileRubber') * 10)
     .EUt(VA[MV])
     .duration(80)
